+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+- Add feature to copy results to clipboard
+
 ## [0.5.0] - 2020-05-18
 
 - Fix command ordering to match documentation